Create a view model and a repository
The view model is for the HomeFragment and the repository is for the User model.

+++ app/src/main/java/dev/sertan/android/harcamatakip/data/model/ExchangeRate.kt
@@ -2,9 +2,9 @@ package dev.sertan.android.harcamatakip.data.model
 
 data class ExchangeRate(
     val base: String = Currency.LIRA.code,
-    val rates: Map<String, Float> = mapOf(
-        Currency.EURO.code to 1f,
-        Currency.POUND.code to 1f,
-        Currency.DOLLAR.code to 1f
+    val rates: Map<String, Double> = mapOf(
+        Currency.EURO.code to 1.0,
+        Currency.POUND.code to 1.0,
+        Currency.DOLLAR.code to 1.0
     )
 )
